git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cinvariants-export@38585 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/octagon/values/OctagonIntValue.java
@@ -67,7 +67,7 @@ public class OctagonIntValue extends OctagonNumericValue<Long> {
 
   @Override
   public OctagonNumericValue add(OctagonNumericValue pVal) {
-    return pVal.add(value.longValue());
+    return pVal.add(value);
   }
 
   @Override
